Move debug logger back to original location and replace usage with class name loggers
