+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -22,6 +22,8 @@ import java.util.stream.Collectors;
 import javax.net.ssl.SSLSocketFactory;
 
 import org.immutables.value.Value;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
@@ -75,9 +77,9 @@ import com.palantir.lock.impl.LockServiceImpl;
 import com.palantir.remoting.ssl.SslConfiguration;
 import com.palantir.remoting.ssl.SslSocketFactories;
 import com.palantir.timestamp.TimestampService;
-import com.palantir.util.DebugLogger;
 
 public final class TransactionManagers {
+    private static final Logger log = LoggerFactory.getLogger(TransactionManagers.class);
     private static final ServiceLoader<AtlasDbFactory> loader = ServiceLoader.load(AtlasDbFactory.class);
     public static final LockClient LOCK_CLIENT = LockClient.of("atlas instance");
 
@@ -106,7 +108,7 @@ public final class TransactionManagers {
             Set<Schema> schemas,
             Environment env,
             boolean allowHiddenTableAccess) {
-        DebugLogger.logger.info("Called TransactionManagers.create on thread {}. This should only happen once.",
+        log.info("Called TransactionManagers.create on thread {}. This should only happen once.",
                 Thread.currentThread().getName());
         return create(config, schemas, env, LockServerOptions.DEFAULT, allowHiddenTableAccess);
     }
